refactor: dedupe buildFacilityEvent, tighten EnrollResidentModal cast
