+++ frontend/src/pages/class-detail/SessionsTabModals.tsx
@@ -1,6 +1,4 @@
 import { Class } from '@/types/program';
-import { FacilityProgramClassEvent, ProgramClassEvent } from '@/types/events';
-import { SelectedClassStatus } from '@/types/attendance';
 import { RescheduleSessionModal } from './RescheduleSessionModal';
 import { CancelEventModal } from '@/components/schedule/CancelEventModal';
 import {
@@ -14,53 +12,11 @@ import {
 import { ChangeRoomModal, ChangeRoomSession } from './ChangeRoomModal';
 import { SessionDetailSheet } from '@/components/schedule/SessionDetailSheet';
 import {
-    findActiveOverride,
+    buildFacilityEvent,
     getSessionChangeInfo,
     type SessionDisplay
 } from './session-utils';
 import { getInstructorName } from '@/lib/formatters';
-
-function buildFacilityEvent(
-    session: SessionDisplay,
-    classId: number,
-    classEvents: ProgramClassEvent[]
-): FacilityProgramClassEvent {
-    const eventId = session.instance.event_id ?? session.instance.id;
-    const backingEvent = classEvents.find((e) => e.id === eventId) ?? classEvents[0];
-    const activeOverride = findActiveOverride(classEvents, session.instance.date);
-    const parts = session.instance.class_time.split('-');
-    const [sh = 0, sm = 0] = (parts[0] ?? '').split(':').map(Number);
-    const [eh = 0, em = 0] = (parts[1] ?? '').split(':').map(Number);
-    const start = new Date(session.dateObj);
-    start.setHours(sh, sm, 0, 0);
-    const end = new Date(session.dateObj);
-    end.setHours(eh, em, 0, 0);
-    return {
-        id: eventId,
-        class_id: classId,
-        duration: backingEvent?.duration ?? '',
-        room_id: activeOverride?.room_id ?? backingEvent?.room_id ?? 0,
-        recurrence_rule: backingEvent?.recurrence_rule ?? '',
-        is_cancelled: session.instance.is_cancelled,
-        instructor_id: activeOverride?.instructor_id ?? backingEvent?.instructor_id ?? null,
-        overrides: backingEvent?.overrides ?? [],
-        reason: null,
-        start,
-        end,
-        is_override: !!activeOverride || !!session.instance.override_id,
-        override_id: activeOverride?.id ?? session.instance.override_id ?? 0,
-        linked_override_event: null,
-        room: '',
-        instructor_name: '',
-        program_id: 0,
-        program_name: '',
-        title: '',
-        enrolled_users: '',
-        frequency: '',
-        credit_types: '',
-        class_status: SelectedClassStatus.Scheduled
-    };
-}
 
 interface SessionsTabModalsProps {
     cls: Class;
